Creates a BaseJobComponent with the boilerplate question-fetching logic that other Components should inherit from. This limits the amount of code that needs to be copied between components.
